tests/baseline/protocol-1.2/inspector/console/console-error-on-call-frame-expected.txt

CONSOLE ERROR: line 1: 42
Tests that console.error does not throw exception when executed in console on call frame.

Set timer for test function.
Script execution paused.
Script execution resumed.
console.error(42)
VM:1 42
(anonymous) @ VM:1
evaluate
testFunction @ console-error-on-call-frame.html:10
undefined
